Move fields to common files
