+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/util/JudgeDecisionHelper.java
@@ -5,8 +5,12 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields;
 
 import java.util.Map;
+
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NO_VALUE;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
 
 @Component
 @Slf4j
@@ -14,40 +18,35 @@ import java.util.Map;
 public class JudgeDecisionHelper {
 
     @NoArgsConstructor(access = AccessLevel.PRIVATE)
-    public static class JudgeFields {
-        public static final String JUDGE_COSTS_CLAIM_GRANTED = "JudgeCostsClaimGranted";
-    }
 
     public static class JudgeDecisions {
-        public static final String YES_VALUE = "Yes";
-        public static final String NO_VALUE = "No";
         public static final String ADJOURN_VALUE = "Adjourn";
     }
 
     public static boolean isJudgeCostClaimGranted(Map<String, Object> caseData) {
-        return isGranted(caseData, JudgeFields.JUDGE_COSTS_CLAIM_GRANTED);
+        return isGranted(caseData, CcdFields.JUDGE_COSTS_CLAIM_GRANTED);
     }
 
     public static boolean isJudgeCostClaimRejected(Map<String, Object> caseData) {
-        return isRejected(caseData, JudgeFields.JUDGE_COSTS_CLAIM_GRANTED);
+        return isRejected(caseData, CcdFields.JUDGE_COSTS_CLAIM_GRANTED);
     }
 
     public static boolean isJudgeCostClaimAdjourned(Map<String, Object> caseData) {
-        return isAdjourned(caseData, JudgeFields.JUDGE_COSTS_CLAIM_GRANTED);
+        return isAdjourned(caseData, CcdFields.JUDGE_COSTS_CLAIM_GRANTED);
     }
 
     public static boolean isJudgeCostClaimEmpty(Map<String, Object> caseData) {
-        String judgeDecision = (String) caseData.get(JudgeFields.JUDGE_COSTS_CLAIM_GRANTED);
+        String judgeDecision = (String) caseData.get(CcdFields.JUDGE_COSTS_CLAIM_GRANTED);
 
         return Strings.isNullOrEmpty(judgeDecision);
     }
 
     private static boolean isGranted(Map<String, Object> caseData, String field) {
-        return JudgeDecisions.YES_VALUE.equalsIgnoreCase((String) caseData.get(field));
+        return YES_VALUE.equalsIgnoreCase((String) caseData.get(field));
     }
 
     private static boolean isRejected(Map<String, Object> caseData, String field) {
-        return JudgeDecisions.NO_VALUE.equalsIgnoreCase((String) caseData.get(field));
+        return NO_VALUE.equalsIgnoreCase((String) caseData.get(field));
     }
 
     private static boolean isAdjourned(Map<String, Object> caseData, String field) {
